+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 1.5.1
+Mirai console 1.0-RC
+
+# 1.5.0
+LuckPerms 5.1 -> 5.2
+
+# 1.4.0
+mirai-console-1.0-RC-dev-28
+
+# 1.3.0
+Mirai-Console M4
+
 # 1.0.1
 
 - 修复权限判断错误.